Update cohort_mode_workspace_dashboard.md.tmpl (#612)
* Update cohort_mode_workspace_dashboard.md.tmpl

Add extra information after the description of Module 10 to guide the user as to which workflow versions to run. This clarifies which versions to use based on the batches to ensure the workflows are run as intended.

* Update inputs/templates/terra_workspaces/cohort_mode/cohort_mode_workspace_dashboard.md.tmpl

I see what you're saying. I was going off of how they were referred to in the dashboard prior. Should that be changed as well?

Co-authored-by: epiercehoffman <epierceh@broadinstitute.org>

---------

Co-authored-by: epiercehoffman <epierceh@broadinstitute.org>