Use `communicate()` with Python subprocess
Using `poll()` on a Popen object will return `None` if the child process
has not terminated and that can cause trouble with the
`get_command_output()` function in the Docker building script. The
function assumes that `poll()` always returns the exit code of the child
process and compares it to 0 to check if the child process executed
successfully. If the child process has not terminated by the time
`poll()` is called, then the comparison will always be false and the
child process will appear to have failed.