Update representation of BNDs to be Pysam compatible (#835)
This PR includes code to refactor VCFs passed through GATK-SV to align it to Pysam's upgraded versions, which do not allow variants to have an END position that is less than its start position. As it stands, VCFs in GATK-SV only use the correct END format starting from the outputs of 13-ResolveComplexVariants.

We hence ensure that all VCFs passed through GATK-SV represent the END2 field to be the true end position on the second chromosome, whereas the END field is set to pos. All downstream scripts that consume such VCFs are made to interpret BNDs with appropriate logic based on this updated END2 tag.